requirements.txt
scipy==1.6.2
numpy==1.19.2
matplotlib==3.3.4
opencv_python==4.2.0.34
torch==1.8.0
torchvision==0.9.0
tqdm==4.55.1
Pillow==9.2.0
h5py==2.10.0
